chore: more component organization
