+++ src/components/homepage/ProjectIcon.tsx
